+++ CustomCADs.Tests/ServicesTests/CadTests/BaseCadsTests.cs
@@ -41,14 +41,14 @@ namespace CustomCADs.Tests.ServicesTests.CadTests
         ];
         protected CadModel[] cads =
         [
-            new() { Id = 1, Name = "Cad1", CadExtension = "a", ImageExtension = "abcdefgh", CategoryId = 1, Price = 1m, Status = CadStatus.Unchecked, CreationDate = DateTime.Now.AddDays(-1), Coords = new int[3], PanCoords = new int[3] },
-            new() { Id = 2, Name = "Cad2", CadExtension = "ab", ImageExtension = "abcdefg", CategoryId = 2, Price = 2m, Status = CadStatus.Unchecked, CreationDate = DateTime.Now.AddDays(-2), Coords = new int[3], PanCoords = new int[3] },
-            new() { Id = 3, Name = "Cad3", CadExtension = "abc", ImageExtension = "abcdef", CategoryId = 3, Price = 3m, Status = CadStatus.Validated, CreationDate = DateTime.Now.AddDays(-3), Coords = new int[3], PanCoords = new int[3] },
-            new() { Id = 4, Name = "Cad4", CadExtension = "abcd", ImageExtension = "abcde", CategoryId = 4, Price = 4m, Status = CadStatus.Validated, CreationDate = DateTime.Now.AddDays(-4), Coords = new int[3], PanCoords = new int[3] },
-            new() { Id = 5, Name = "Cad5", CadExtension = "abcde", ImageExtension = "abcd", CategoryId = 5, Price = 5m, Status = CadStatus.Reported, CreationDate = DateTime.Now.AddDays(-5), Coords = new int[3], PanCoords = new int[3] },
-            new() { Id = 6, Name = "Cad6", CadExtension = "abcdef", ImageExtension = "abc", CategoryId = 6, Price = 6m, Status = CadStatus.Reported, CreationDate = DateTime.Now.AddDays(-6), Coords = new int[3], PanCoords = new int[3] },
-            new() { Id = 7, Name = "Cad7", CadExtension = "abcdefg", ImageExtension = "ab", CategoryId = 7, Price = 7m, Status = CadStatus.Banned, CreationDate = DateTime.Now.AddDays(-7), Coords = new int[3], PanCoords = new int[3] },
-            new() { Id = 8, Name = "Cad8", CadExtension = "abcdefgh", ImageExtension = "a", CategoryId = 8, Price = 8m, Status = CadStatus.Banned, CreationDate = DateTime.Now.AddDays(-8), Coords = new int[3], PanCoords = new int[3] },
+            new() { Id = 1, Name = "Cad1", CadExtension = "a", ImageExtension = "abcdefgh", CategoryId = 1, Price = 1m, Status = CadStatus.Unchecked, CreationDate = DateTime.Now.AddDays(-1), Coords = new double[3], PanCoords = new double[3] },
+            new() { Id = 2, Name = "Cad2", CadExtension = "ab", ImageExtension = "abcdefg", CategoryId = 2, Price = 2m, Status = CadStatus.Unchecked, CreationDate = DateTime.Now.AddDays(-2), Coords = new double[3], PanCoords = new double[3] },
+            new() { Id = 3, Name = "Cad3", CadExtension = "abc", ImageExtension = "abcdef", CategoryId = 3, Price = 3m, Status = CadStatus.Validated, CreationDate = DateTime.Now.AddDays(-3), Coords = new double[3], PanCoords = new double[3] },
+            new() { Id = 4, Name = "Cad4", CadExtension = "abcd", ImageExtension = "abcde", CategoryId = 4, Price = 4m, Status = CadStatus.Validated, CreationDate = DateTime.Now.AddDays(-4), Coords = new double[3], PanCoords = new double[3] },
+            new() { Id = 5, Name = "Cad5", CadExtension = "abcde", ImageExtension = "abcd", CategoryId = 5, Price = 5m, Status = CadStatus.Reported, CreationDate = DateTime.Now.AddDays(-5), Coords = new double[3], PanCoords = new double[3] },
+            new() { Id = 6, Name = "Cad6", CadExtension = "abcdef", ImageExtension = "abc", CategoryId = 6, Price = 6m, Status = CadStatus.Reported, CreationDate = DateTime.Now.AddDays(-6), Coords = new double[3], PanCoords = new double[3] },
+            new() { Id = 7, Name = "Cad7", CadExtension = "abcdefg", ImageExtension = "ab", CategoryId = 7, Price = 7m, Status = CadStatus.Banned, CreationDate = DateTime.Now.AddDays(-7), Coords = new double[3], PanCoords = new double[3] },
+            new() { Id = 8, Name = "Cad8", CadExtension = "abcdefgh", ImageExtension = "a", CategoryId = 8, Price = 8m, Status = CadStatus.Banned, CreationDate = DateTime.Now.AddDays(-8), Coords = new double[3], PanCoords = new double[3] },
         ];
 
         [OneTimeSetUp]
